+++ poutyne/framework/metrics/epoch_metrics/base.py
@@ -35,3 +35,11 @@ class EpochMetric(ABC, nn.Module):
         that a new epoch can be done.
         """
         pass
+
+    # For compatibility with torchmetrics.
+    def update(self, y_pred, y_true) -> None:
+        return self.forward(y_pred, y_true)
+
+    # For compatibility with torchmetrics.
+    def compute(self):
+        return self.get_metric()