+++ poutyne/framework/metrics/utils.py
@@ -1,5 +1,6 @@
 import re
 from collections import Counter
+from typing import List, Tuple
 
 # From https://stackoverflow.com/a/1176023
 pattern1 = re.compile(r'(.)([A-Z][a-z]+)')
@@ -33,21 +34,23 @@ def flatten_metric_names(metric_names):
     return [name for names in metric_names for name in to_list(names)]
 
 
-def rename_doubles(batch_metrics_names, epoch_metrics_names):
-    counts = Counter(flatten_metric_names(batch_metrics_names + epoch_metrics_names))
-    numbering = Counter()
-    batch_metrics_names = rename_doubles_from_counts(batch_metrics_names, counts, numbering)
-    epoch_metrics_names = rename_doubles_from_counts(epoch_metrics_names, counts, numbering)
-    return batch_metrics_names, epoch_metrics_names
+def rename_doubles(batch_metrics_names: List, epoch_metrics_names: List, torch_metrics_names: List) -> Tuple:
+    metrics_names = rename_doubles_from_list(batch_metrics_names + epoch_metrics_names + torch_metrics_names)
+    batch_metrics_names = metrics_names[: len(batch_metrics_names)]
+    epoch_metrics_names = metrics_names[len(batch_metrics_names) : len(batch_metrics_names) + len(epoch_metrics_names)]
+    torch_metrics_names = metrics_names[len(batch_metrics_names) + len(epoch_metrics_names) :]
+    return batch_metrics_names, epoch_metrics_names, torch_metrics_names
 
 
-def rename_doubles_from_counts(metric_names, counts, numbering):
+def rename_doubles_from_list(metric_names: List) -> List:
     """
     This function takes a list in the format `['a', ['b', 'a'], 'c', 'a', 'c']`
     and returns a list where each double is added a number so that there are no
     more doubles in the list: `['a1', ['b', 'a2'], 'c1', 'a3', 'c2']`. It does so
     using the provided counts and using the numbering Counter object.
     """
+    counts = Counter(flatten_metric_names(metric_names))
+    numbering = Counter()
 
     def get_name(name):
         if counts[name] > 1: