+++ tests/framework/metrics/test_metrics_model_integration.py
@@ -9,8 +9,21 @@ from unittest.mock import ANY
 import numpy as np
 import torch
 import torch.nn as nn
+import torchmetrics
 
 from poutyne import Model, EpochMetric, rename_doubles, register_epoch_metric_class, unregister_epoch_metric
+
+
+class MyConstTorchMetric(torchmetrics.Metric):
+    def __init__(self, value=0):
+        super().__init__()
+        self.value = value
+
+    def update(self, preds: torch.Tensor, target: torch.Tensor):
+        pass
+
+    def compute(self):
+        return self.value
 
 
 class ConstEpochMetric(EpochMetric):
@@ -51,6 +64,7 @@ def get_const_batch_metric(value):
 
 
 class MetricsModelIntegrationTest(unittest.TestCase):
+    # pylint: disable=too-many-public-methods
     epochs = 2
     steps_per_epoch = 3
     batch_size = 10
@@ -220,6 +234,59 @@ class MetricsModelIntegrationTest(unittest.TestCase):
         self._test_history(model, ['unique_some', 'something'], [0.0, 0.0])
         unregister_epoch_metric(names)
 
+    def test_torchmetrics_handling(self):
+        expected_names = ['r2_score', 'spearman_corr_coef']
+        model = Model(
+            self.pytorch_network,
+            self.optimizer,
+            self.loss_function,
+            torch_metrics=[torchmetrics.R2Score(), torchmetrics.SpearmanCorrCoef()],
+        )
+        self._test_history(model, expected_names, [ANY, ANY])
+
+    def test_repeated_torch_metrics_handling(self):
+        expected_names = ['my_const_torch_metric1', 'my_const_torch_metric2']
+        model = Model(
+            self.pytorch_network,
+            self.optimizer,
+            self.loss_function,
+            torch_metrics=[MyConstTorchMetric(1), MyConstTorchMetric(2)],
+        )
+        self._test_history(model, expected_names, [1, 2])
+
+    def test_repeated_torch_metrics_handling_with_different_names(self):
+        expected_names = ['metric1', 'metric2']
+        model = Model(
+            self.pytorch_network,
+            self.optimizer,
+            self.loss_function,
+            torch_metrics=[('metric1', MyConstTorchMetric(1)), ('metric2', MyConstTorchMetric(2))],
+        )
+        self._test_history(model, expected_names, [1, 2])
+
+    def test_use_torch_metrics_as_epoch_metrics(self):
+        expected_names = ['metric1', 'metric2']
+        model = Model(
+            self.pytorch_network,
+            self.optimizer,
+            self.loss_function,
+            epoch_metrics=[('metric1', MyConstTorchMetric(1)), ('metric2', MyConstTorchMetric(2))],
+        )
+        self._test_history(model, expected_names, [1, 2])
+
+    @skipIf(not torch.cuda.is_available(), "no gpu available")
+    def test_torch_metrics_on_gpu(self):
+        with torch.cuda.device(MetricsModelIntegrationTest.cuda_device):
+            expected_names = ['r2_score', 'spearman_corr_coef']
+            model = Model(
+                self.pytorch_network,
+                self.optimizer,
+                self.loss_function,
+                torch_metrics=[torchmetrics.R2Score(), torchmetrics.SpearmanCorrCoef()],
+            )
+            model.cuda()
+            self._test_history(model, expected_names, [ANY, ANY])
+
     def _test_history(self, model, names, values):
         history = model.fit(
             self.train_x,
@@ -262,63 +329,72 @@ class MetricsModelIntegrationTest(unittest.TestCase):
 
 class MetricsRenamingTest(unittest.TestCase):
     def test_batch_metrics(self):
-        actual = rename_doubles(['a', 'a'], [])
-        expected = ['a1', 'a2'], []
+        actual = rename_doubles(['a', 'a'], [], [])
+        expected = ['a1', 'a2'], [], []
         self.assertEqual(expected, actual)
 
-        actual = rename_doubles(['a', 'b', 'a', 'a', 'c', 'd'], [])
-        expected = ['a1', 'b', 'a2', 'a3', 'c', 'd'], []
+        actual = rename_doubles(['a', 'b', 'a', 'a', 'c', 'd'], [], [])
+        expected = ['a1', 'b', 'a2', 'a3', 'c', 'd'], [], []
         self.assertEqual(expected, actual)
 
     def test_epoch_metrics(self):
-        actual = rename_doubles([], ['a', 'a'])
-        expected = [], ['a1', 'a2']
+        actual = rename_doubles([], ['a', 'a'], [])
+        expected = [], ['a1', 'a2'], []
         self.assertEqual(expected, actual)
 
-        actual = rename_doubles([], ['a', 'b', 'a', 'a', 'c', 'd'])
-        expected = [], ['a1', 'b', 'a2', 'a3', 'c', 'd']
+        actual = rename_doubles([], ['a', 'b', 'a', 'a', 'c', 'd'], [])
+        expected = [], ['a1', 'b', 'a2', 'a3', 'c', 'd'], []
         self.assertEqual(expected, actual)
 
-    def test_batch_epoch_metrics(self):
-        actual = rename_doubles(['a', 'b', 'c'], ['d', 'a', 'e', 'a'])
-        expected = ['a1', 'b', 'c'], ['d', 'a2', 'e', 'a3']
+    def test_torch_metrics(self):
+        actual = rename_doubles([], [], ['a', 'a'])
+        expected = [], [], ['a1', 'a2']
         self.assertEqual(expected, actual)
 
-        actual = rename_doubles(['a', 'b', 'c', 'b'], ['d', 'a', 'e', 'a', 'e'])
-        expected = ['a1', 'b1', 'c', 'b2'], ['d', 'a2', 'e1', 'a3', 'e2']
+        actual = rename_doubles([], [], ['a', 'b', 'a', 'a', 'c', 'd'])
+        expected = [], [], ['a1', 'b', 'a2', 'a3', 'c', 'd']
         self.assertEqual(expected, actual)
 
-        actual = rename_doubles(['a', 'b', 'c'], ['d', 'a', 'e', 'a', 'e'])
-        expected = ['a1', 'b', 'c'], ['d', 'a2', 'e1', 'a3', 'e2']
+    def test_batch_epoch_torch_metrics(self):
+        actual = rename_doubles(['a', 'b', 'c'], ['d', 'a', 'e', 'a'], ['f', 'a', 'g'])
+        expected = ['a1', 'b', 'c'], ['d', 'a2', 'e', 'a3'], ['f', 'a4', 'g']
+        self.assertEqual(expected, actual)
+
+        actual = rename_doubles(['a', 'b', 'c', 'b'], ['d', 'a', 'e', 'a', 'e'], ['f', 'a'])
+        expected = ['a1', 'b1', 'c', 'b2'], ['d', 'a2', 'e1', 'a3', 'e2'], ['f', 'a4']
+        self.assertEqual(expected, actual)
+
+        actual = rename_doubles(['a', 'b', 'c'], ['d', 'a', 'e', 'a', 'e'], ['f', 'a'])
+        expected = ['a1', 'b', 'c'], ['d', 'a2', 'e1', 'a3', 'e2'], ['f', 'a4']
         self.assertEqual(expected, actual)
 
     def test_nested_batch_metrics(self):
-        actual = rename_doubles([['a', 'a']], [])
-        expected = [['a1', 'a2']], []
+        actual = rename_doubles([['a', 'a']], [], [])
+        expected = [['a1', 'a2']], [], []
         self.assertEqual(expected, actual)
 
-        actual = rename_doubles([['a', 'b'], 'b'], [])
-        expected = [['a', 'b1'], 'b2'], []
+        actual = rename_doubles([['a', 'b'], 'b'], [], [])
+        expected = [['a', 'b1'], 'b2'], [], []
         self.assertEqual(expected, actual)
 
     def test_nested_epoch_metrics(self):
-        actual = rename_doubles([], [['a', 'a']])
-        expected = [], [['a1', 'a2']]
+        actual = rename_doubles([], [['a', 'a']], [])
+        expected = [], [['a1', 'a2']], []
         self.assertEqual(expected, actual)
 
-        actual = rename_doubles([], [['a', 'b'], 'b'])
-        expected = [], [['a', 'b1'], 'b2']
+        actual = rename_doubles([], [['a', 'b'], 'b'], [])
+        expected = [], [['a', 'b1'], 'b2'], []
         self.assertEqual(expected, actual)
 
     def test_nested_batch_epoch_metrics(self):
-        actual = rename_doubles([['a']], [['a']])
-        expected = [['a1']], [['a2']]
+        actual = rename_doubles([['a']], [['a']], [])
+        expected = [['a1']], [['a2']], []
         self.assertEqual(expected, actual)
 
-        actual = rename_doubles([['a', 'b']], [['c', 'a']])
-        expected = [['a1', 'b']], [['c', 'a2']]
+        actual = rename_doubles([['a', 'b']], [['c', 'a']], [])
+        expected = [['a1', 'b']], [['c', 'a2']], []
         self.assertEqual(expected, actual)
 
-        actual = rename_doubles(['a', 'b', ['b', 'a'], 'c', 'a', 'd'], ['e', ['c', 'a'], 'f', 'a'])
-        expected = ['a1', 'b1', ['b2', 'a2'], 'c1', 'a3', 'd'], ['e', ['c2', 'a4'], 'f', 'a5']
+        actual = rename_doubles(['a', 'b', ['b', 'a'], 'c', 'a', 'd'], ['e', ['c', 'a'], 'f', 'a'], [])
+        expected = ['a1', 'b1', ['b2', 'a2'], 'c1', 'a3', 'd'], ['e', ['c2', 'a4'], 'f', 'a5'], []
         self.assertEqual(expected, actual)