+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # v1.x.x
+* Add support for [TorchMetrics](https://torchmetrics.readthedocs.io/) metrics.
 * [`Experiment`](https://poutyne.org/experiment.html#poutyne.Experiment) is now an alias for
   [`ModelBundle`](https://poutyne.org/experiment.html#poutyne.ModelBundle), a class quite similar to `Experiment` except
   that it allows to instantiate an "Experiment" from a Poutyne Model or a network.